Release edge
