--- edge/build.json
@@ -1,9 +1,0 @@
-{
-    "build_from": {
-      "aarch64": "homeassistant/aarch64-base-python:3.9-alpine3.13",
-      "amd64": "homeassistant/amd64-base-python:3.9-alpine3.13",
-      "armhf": "homeassistant/armhf-base-python:3.9-alpine3.13",
-      "armv7": "homeassistant/armv7-base-python:3.9-alpine3.13",
-      "i386": "homeassistant/i386-base-python:3.9-alpine3.13"
-    }
-  }
