--- edge/config.json
@@ -1,62 +1,0 @@
-{
-  "name": "HoyMiles Solar Gateway Edge Addon",
-  "version": "1.3.1",
-  "slug": "hoymiles_solardata_edge",
-  "description": "Read Hoymiles Gateway Solar Data Edge Version.",
-  "arch": [
-    "armhf",
-    "armv7",
-    "aarch64",
-    "amd64",
-    "i386"
-  ],
-  "url": "https://github.com/dmslabsbr/hoymiles/tree/master/edge",
-  "startup": "application",
-  "boot": "auto",
-  "init": false,
-  "hassio_api": false,
-  "homeassistant_api": false,
-  "hassio_role": "default",
-  "devicetree": false,
-  "services": [
-    "mqtt:need"
-  ],
-  "ingress": false,
-  "ingress_port": 5000,
-  "host_network": true,
-  "panel_icon": "mdi:solar-power",
-  "panel_title": "HoyMiles Solar Data Gateway Edge version.",
-  "panel_admin": false,
-  "options": {
-    "HOYMILES_USER": "HOYMILES_USER",
-    "HOYMILES_PASSWORD": "HOYMILES_PASSWORD",
-    "HOYMILES_PLANT_ID": "123",
-    "Read_meter_data": true,
-    "External_MQTT_Server": false,
-    "MQTT_Host": "Internal - Change here if you use another server",
-    "MQTT_User": "Automatically obtained from Home Assistant - Change here if you use another server",
-    "MQTT_Pass": "Automatically obtained from Home Assistant - Change here if you use another server",
-    "MQTT_TLS": false,
-    "MQTT_TLS_PORT": "8883",
-    "DEVELOPERS_MODE": false,
-    "LOG_LEVEL": "INFO",
-    "LOG_TO_FILE": false,
-    "FILE_PATH": "hoymiles.log"
-  },
-  "schema": {
-    "HOYMILES_USER": "str",
-    "HOYMILES_PASSWORD": "password",
-    "HOYMILES_PLANT_ID": "str",
-    "Read_meter_data": "bool",
-    "External_MQTT_Server": "bool",
-    "MQTT_Host": "str?",
-    "MQTT_User": "str?",
-    "MQTT_Pass": "password?",
-    "MQTT_TLS": "bool",
-    "MQTT_TLS_PORT": "port",
-    "DEVELOPERS_MODE": "bool",
-    "LOG_LEVEL": "list(DEBUG|INFO|WARNING|ERROR|CRITICAL)?",
-    "LOG_TO_FILE": "bool",
-    "FILE_PATH": "str?"
-  }
-}
