--- edge/jsons/dtu.json
@@ -1,15 +1,0 @@
-{
-    "binary_sensor": {
-        "connect": {
-            "name": "connect",
-            "device_class": "connectivity",
-            "icon": "mdi:solar-panel"
-        },
-        "todos": {
-            "name": "",
-            "device_class": "",
-            "icon": ""
-        }
-    }
-
-}
