--- edge/jsons/micro_codes
@@ -1,202 +1,0 @@
-TC function enabled:10074
-TC function enabled:74
-Reset:3001
-Reset:3002
-Reset:3003
-Reset:3004
-Grid transient fluctuation:5160
-Error Code 14:5141
-Error Code 14:5142
-Error Code 14:5143
-Error Code 14:5144
-Over temperature protection:5070
-Locked by anti-theft lock:9000
-Locked by remote control:8320
-Shut down by remote control:8310
-Error Code 12:5120
-Firmware error:5200
-Error Code 06:5060
-Grid overvoltage (10 min mean value):5150
-Island detected:5100
-Standby:2001
-Standby:2002
-Standby:2003
-Standby:2004
-Grid overvoltage/undervoltage:5080
-Grid overfrequency/underfrequency:5090
-GFDI failure:5110
-PV1 input overvoltage/undervoltage:5051
-PV2 input overvoltage/undervoltage:5052
-PV3 input overvoltage/undervoltage:5053
-PV4 input overvoltage/undervoltage:5054
-Error Code 01:5011
-Error Code 01:5012
-Error Code 01:5013
-Error Code 01:5014
-Error Code 02:5020
-Error Code 04:5041
-Error Code 04:5042
-Error Code 04:5043
-Error Code 04:5044
-Offline:130
-Error Code 01:5511
-Error Code 01:5512
-Error Code 01:5513
-Error Code 01:5514
-Error Code 02:5520
-Reset:1
-Time calibration:2
-EEPROM reading and writing error:3
-Offline:4
-Grid voltage surge:11
-Grid voltage sharp drop:12
-Grid voltage spike:13
-rid phase mutation:14
-Grid zero-crossing distortion:15
-INV overcurrent/overvoltage:36
-FB overcurrent:46
-FB overcurrent:47
-FB clamp overvoltage:48
-FB clamp overvoltage:49
-FB overcurrent:50
-FB clamp overvoltage:51
-FB overcurrent:52
-FB clamp overvoltage:53
-Calibration parameter error:61
-System configuration parameter error:62
-Abnormal power generation data:63
-VW function enabled:71
-FW function enabled:72
-TW function enabled:73
-PV module in suspected shadow:95
-PV module in suspected shadow:96
-PV module in suspected shadow:97
-PV module in suspected shadow:98
-Over temperature protection:121
-Locked by anti-theft lock:122
-Locked by remote control:123
-Shut down by remote control:124
-Grid configuration parameter error:125
-Firmware error:127
-Software Error Code 128:128
-Software Error Code 129:129
-Grid overvoltage:141
-Grid overvoltage (10 min mean value):142
-Grid undervoltage:143
-Grid overfrequency:144
-Grid underfrequency:145
-High RoCoF:146
-No voltage on grid port:147
-Grid disconnection:148
-Island detected:149
-DCI exceeded:150
-Severe grid wave distortion:151
-Abnormal insulation impedance:181
-Abnormal grounding:182
-MPPT-A input overvoltage:205
-MPPT-B input overvoltage:206
-MPPT-A input undervoltage:207
-MPPT-B input undervoltage:208
-PV1 no input:209
-PV2 no input:210
-PV3 no input:211
-PV4 no input:212
-PV1 & PV2 abnormal wiring:213
-PV3 & PV4 abnormal wiring:214
-PV1 input overvoltage:215
-PV1 input undervoltage:216
-PV2 input overvoltage:217
-PV2 input undervoltage:218
-PV3 input overvoltage:219
-PV3 input undervoltage:220
-PV4 input overvoltage:221
-PV4 input undervoltage:222
-Hardware Error Code 301:301
-Hardware Error Code 302:302
-Hardware Error Code 303:303
-Hardware Error Code 304:304
-Hardware Error Code 305:305
-Hardware Error Code 306:306
-Hardware Error Code 307:307
-Hardware Error Code 308:308
-Hardware Error Code 309:309
-Hardware Error Code 310:310
-Hardware Error Code 311:311
-Hardware Error Code 312:312
-Hardware Error Code 313:313
-Hardware Error Code 314:314
-EEPROM reading and writing error:10003
-Offline:10004
-Grid voltage surge:10011
-Grid voltage sharp drop:10012
-Grid voltage spike:10013
-Grid phase mutation:10014
-Grid zero-crossing distortion:10015
-INV overcurrent/overvoltage:10036
-FB overcurrent:10046
-FB overcurrent:10047
-FB clamp overvoltage:10048
-FB clamp overvoltage:10049
-FB overcurrent:10050
-FB clamp overvoltage:10051
-Calibration parameter error:10061
-System configuration parameter error:10062
-Abnormal power generation data:10063
-VW function enabled:10071
-FW function enabled:10072
-TW function enabled:10073
-PV module in suspected shadow:10095
-PV module in suspected shadow:10096
-PV module in suspected shadow:10097
-PV module in suspected shadow (invalid):10098
-PV module in suspected shadow (invalid):10099
-PV module in suspected shadow (invalid):10100
-Over temperature protection:10121
-Locked by anti-theft lock:10122
-Locked by remote control:10123
-Shut down by remote control:10124
-Grid configuration parameter error:10125
-Firmware error:10127
-Software Error Code 128:10128
-Software Error Code 129:10129
-Offline:10130
-Grid overvoltage:10141
-Grid overvoltage (10 min mean value):10142
-Grid undervoltage:10143
-Grid overfrequency:10144
-Grid underfrequency:10145
-High RoCoF:10146
-No voltage on grid port:10147
-Grid disconnection:10148
-Island detected:10149
-DCI exceeded:10150
-Abnormal grid phase:10152
-Abnormal insulation impedance:10181
-Abnormal grounding:10182
-MPPT-A input overvoltage:10205
-MPPT-B input overvoltage:10206
-MPPT-A input undervoltage:10207
-MPPT-B input undervoltage:10208
-PV1 no input:10209
-PV2 no input:10210
-PV3 no input:10211
-PV4 no input:10212
-PV1 & PV2 abnormal wiring:10213
-PV3 & PV4 abnormal wiring:10214
-MPPT-C input overvoltage:10215
-MPPT-C input undervoltage:10216
-PV5 no input:10217
-PV6 no input:10218
-PV5 & PV6 abnormal wiring:10219
-Hardware Error Code 301:10301
-Hardware Error Code 302:10302
-Hardware Error Code 303:10303
-Hardware Error Code 304:10304
-Hardware Error Code 305:10305
-Hardware Error Code 306:10306
-Hardware Error Code 307:10307
-Hardware Error Code 308:10308
-Hardware Error Code 309:10309
-Hardware Error Code 310:10310
-Hardware Error Code 311:10311
-Abnormal wiring of grid neutral wire:10153
