--- edge/repository.json
@@ -1,5 +1,0 @@
-{
-  "name": "Hoymiles Home Assistant Edge Addon",
-  "url": "https://github.com/dmslabsbr/hoymiles/edge",
-  "maintainer": "Daniel <suport@neoage.com.br>"
-}
