--- edge/templates/base.html
@@ -1,38 +1,0 @@
-<html>
-
-  <style>
-    @media (prefers-color-scheme: dark) {
-      div, body {
-        background: black;
-        color: yellow;
-      }
-    }
-    @media (prefers-color-scheme: light) {
-      div,body {
-        background: white;
-        color: green;
-      }
-    }
-  </style>
-
-<script>
-window.matchMedia('(prefers-color-scheme: dark)').addListener(function (e) {
-    console.log(`changed to ${e.matches ? "dark" : "light"} mode`)
-});
-
-</script>
-
-
-    <head>
-      {% if title %}
-      <title>{{ title }} - Microblog</title>
-      {% else %}
-      <title>Welcome to Microblog</title>
-      {% endif %}
-    </head>
-    <body>
-        <div></div>
-        <hr>
-        {% block content %}{% endblock %}
-    </body>
-</html>
