--- edge/templates/index.html
@@ -1,10 +1,0 @@
-{% extends "base.html" %}
-
-{% block content %}
-    <TABLE>
-    {% for item, valor in itens.items() -%}
-        <TR></TR>
-        <TD><b>{{ item }}:</b></TD><td>{{ valor }}</td>
-    {% endfor %}
-    </TABLE>
-{% endblock %}
