+++ oldStable/Dockerfile
@@ -7,7 +7,7 @@ ENV LANG C.UTF-8
 # erro em amd64-base-python:3.9-alpine3.12
 
 # Install requirements for add-on
-#RUN apk add --no-cache python3 py3-pip
+#RUN apk add --no-cache python3 py3-pip 
 
 RUN python -m pip install --upgrade pip
 RUN pip3 install paho-mqtt==1.6.1
