+++ oldStable/jsons/sensor.json
@@ -1,12 +1,12 @@
 {
-   "capacitor":{
-      "name":"capacitor",
+   "array_size":{
+      "name":"array_size",
       "device_class":"power",
       "icon":"mdi:solar-panel",
       "unit_of_measurement":"W"
    },
-   "capacitor_kW":{
-      "name":"capacitor_kW",
+   "array_size_kW":{
+      "name":"array_size_kW",
       "device_class":"power",
       "icon":"mdi:solar-panel",
       "unit_of_measurement":"kW"
@@ -48,6 +48,13 @@
       "device_class":"power",
       "unit_of_measurement":"W"
    },
+   "real_power_kw":{
+      "name":"real_power_kw",
+      "icon":"mdi:solar-power",
+      "state_class":"measurement",
+      "device_class":"power",
+      "unit_of_measurement":"kW"
+   },   
    "real_power_measurement":{
       "name":"real_power_measurement",
       "icon":"mdi:solar-power",
