+++ oldStable/repository.json
@@ -1,5 +1,5 @@
 {
-  "name": "Hoymiles Home Assistant Addon",
+  "name": "Hoymiles Home Assistant oldStable Addon",
   "url": "https://github.com/dmslabsbr/hoymiles/oldStable",
   "maintainer": "Daniel <suport@neoage.com.br>"
 }
