+++ stable/config.json
@@ -1,8 +1,8 @@
 {
-  "name": "HoyMiles Solar Gateway Stable Addon",
-  "version": "1.1.7",
+  "name": "HoyMiles Solar Gateway stable Addon",
+  "version": "1.4.0",
   "slug": "hoymiles_solardata_stable",
-  "description": "Read Hoymiles Gateway Solar Data Stable Version.",
+  "description": "Read Hoymiles Gateway Solar Data stable Version.",
   "arch": [
     "armhf",
     "armv7",
@@ -25,12 +25,13 @@
   "ingress_port": 5000,
   "host_network": true,
   "panel_icon": "mdi:solar-power",
-  "panel_title": "HoyMiles Solar Data Gateway Stable version.",
+  "panel_title": "HoyMiles Solar Data Gateway stable version.",
   "panel_admin": false,
   "options": {
     "HOYMILES_USER": "HOYMILES_USER",
     "HOYMILES_PASSWORD": "HOYMILES_PASSWORD",
     "HOYMILES_PLANT_ID": "123",
+    "Read_meter_data": true,
     "External_MQTT_Server": false,
     "MQTT_Host": "Internal - Change here if you use another server",
     "MQTT_User": "Automatically obtained from Home Assistant - Change here if you use another server",
@@ -46,6 +47,7 @@
     "HOYMILES_USER": "str",
     "HOYMILES_PASSWORD": "password",
     "HOYMILES_PLANT_ID": "str",
+    "Read_meter_data": "bool",
     "External_MQTT_Server": "bool",
     "MQTT_Host": "str?",
     "MQTT_User": "str?",
