+++ stable/jsons/sensor.json
@@ -1,108 +1,170 @@
 {
-   "array_size":{
-      "name":"array_size",
-      "device_class":"power",
-      "icon":"mdi:solar-panel",
-      "unit_of_measurement":"W"
+   "array_size": {
+      "name": "array_size",
+      "device_class": "power",
+      "icon": "mdi:solar-panel",
+      "unit_of_measurement": "W"
    },
-   "array_size_kW":{
-      "name":"array_size_kW",
-      "device_class":"power",
-      "icon":"mdi:solar-panel",
-      "unit_of_measurement":"kW"
+   "array_size_kW": {
+      "name": "array_size_kW",
+      "device_class": "power",
+      "icon": "mdi:solar-panel",
+      "unit_of_measurement": "kW"
    },
-   "co2_emission_reduction":{
-      "name":"co2_emission_reduction",
-      "icon":"mdi:molecule-co2",
-      "state_class":"total_increasing",
-      "unit_of_measurement":"t"
+   "co2_emission_reduction": {
+      "name": "co2_emission_reduction",
+      "icon": "mdi:molecule-co2",
+      "state_class": "total_increasing",
+      "unit_of_measurement": "t"
    },
-   "is_balance":{
-      "name":"is_balance",
-      "icon":"mdi:crosshairs-question"
+   "is_balance": {
+      "name": "is_balance",
+      "icon": "mdi:crosshairs-question"
    },
-   "is_reflux":{
-      "name":"is_reflux",
-      "icon":"mdi:crosshairs-question"
+   "is_reflux": {
+      "name": "is_reflux",
+      "icon": "mdi:crosshairs-question"
    },
-   "last_data_time":{
-      "device_class":"timestamp",
-      "name":"last_data_time"
+   "last_data_time": {
+      "device_class": "timestamp",
+      "name": "last_data_time"
    },
-   "month_eq":{
-      "name":"month_eq",
-      "device_class":"energy",
-      "state_class":"total_increasing",
-      "unit_of_measurement":"kWh"
+   "month_eq": {
+      "name": "month_eq",
+      "device_class": "energy",
+      "state_class": "total_increasing",
+      "unit_of_measurement": "kWh"
    },
-   "plant_tree":{
-      "name":"plant_tree",
-      "icon":"mdi:tree",
-      "state_class":"total_increasing",
-      "unit_of_measurement":"Trees"
+   "plant_tree": {
+      "name": "plant_tree",
+      "icon": "mdi:tree",
+      "state_class": "total_increasing",
+      "unit_of_measurement": "Trees"
    },
-   "real_power":{
-      "name":"real_power",
-      "icon":"mdi:solar-power",
-      "state_class":"measurement",
-      "device_class":"power",
-      "unit_of_measurement":"W"
+   "real_power": {
+      "name": "real_power",
+      "icon": "mdi:solar-power",
+      "state_class": "measurement",
+      "device_class": "power",
+      "unit_of_measurement": "W"
    },
-   "real_power_kw":{
-      "name":"real_power_kw",
-      "icon":"mdi:solar-power",
-      "state_class":"measurement",
-      "device_class":"power",
-      "unit_of_measurement":"kW"
-   },   
-   "real_power_measurement":{
-      "name":"real_power_measurement",
-      "icon":"mdi:solar-power",
-      "state_class":"measurement",
-      "device_class":"energy",
-      "unit_of_measurement":"Wh"
+   "real_power_kw": {
+      "name": "real_power_kw",
+      "icon": "mdi:solar-power",
+      "state_class": "measurement",
+      "device_class": "power",
+      "unit_of_measurement": "kW"
    },
-   "real_power_total_increasing":{
-      "name":"real_power_total_increasing",
-      "icon":"mdi:solar-power",
-      "state_class":"total",
-      "device_class":"energy",
-      "unit_of_measurement":"Wh"
+   "real_power_measurement": {
+      "name": "real_power_measurement",
+      "icon": "mdi:solar-power",
+      "state_class": "measurement",
+      "device_class": "energy",
+      "unit_of_measurement": "Wh"
    },
-   "reflux_station_data":{
-      "name":"reflux_station_data",
-      "icon":"mdi:crosshairs-question"
+   "real_power_total_increasing": {
+      "name": "real_power_total_increasing",
+      "icon": "mdi:solar-power",
+      "state_class": "total",
+      "device_class": "energy",
+      "unit_of_measurement": "Wh"
    },
-   "today_eq":{
-      "name":"today_eq",
-      "device_class":"energy",
-      "state_class":"total_increasing",
-      "unit_of_measurement":"kWh"
+   "reflux_station_data": {
+      "name": "reflux_station_data.pv_power",
+      "icon": "mdi:crosshairs-question"
    },
-   "today_eq_Wh":{
-      "name":"today_eq_Wh",
-      "device_class":"energy",
-      "state_class":"total_increasing",
-      "unit_of_measurement":"Wh"
+   "today_eq": {
+      "name": "today_eq",
+      "device_class": "energy",
+      "state_class": "total_increasing",
+      "unit_of_measurement": "kWh"
    },
-   "total_eq":{
-      "name":"total_eq",
-      "device_class":"energy",
-      "icon":"mdi:infinity",
-      "state_class":"total_increasing",
-      "unit_of_measurement":"kWh"
+   "today_eq_Wh": {
+      "name": "today_eq_Wh",
+      "device_class": "energy",
+      "state_class": "total_increasing",
+      "unit_of_measurement": "Wh"
    },
-   "power_ratio":{
-      "name":"power_ratio",
-      "device_class":"power_factor",
-      "icon":"mdi:gauge-full",
-      "unit_of_measurement":"%"
+   "total_eq": {
+      "name": "total_eq",
+      "device_class": "energy",
+      "icon": "mdi:infinity",
+      "state_class": "total_increasing",
+      "unit_of_measurement": "kWh"
    },
-   "todos":{ 
-       "name":"",
-       "device_class":"",
-       "state_class":"",
-       "icon":"",
-       "unit_of_measurement":""
+   "power_ratio": {
+      "name": "power_ratio",
+      "device_class": "power_factor",
+      "icon": "mdi:gauge-full",
+      "unit_of_measurement": "%"
+   },
+   "meter_pv_power": {
+      "name": "pv_power",
+      "icon": "mdi:solar-power",
+      "state_class": "measurement",
+      "device_class": "power",
+      "unit_of_measurement": "W"
+   },
+   "meter_grid_power": {
+      "name": "grid_power",
+      "icon": "mdi:solar-power",
+      "state_class": "measurement",
+      "device_class": "power",
+      "unit_of_measurement": "W"
+   },
+   "meter_load_power": {
+      "name": "load_power",
+      "icon": "mdi:solar-power",
+      "state_class": "measurement",
+      "device_class": "power",
+      "unit_of_measurement": "W"
+   },
+   "meter_to_grid": {
+      "name": "meter_b_out_eq",
+      "device_class": "energy",
+      "icon": "mdi:infinity",
+      "state_class": "total_increasing",
+      "unit_of_measurement": "kWh"
+   },
+   "meter_from_grid": {
+      "name": "meter_b_in_eq",
+      "device_class": "energy",
+      "icon": "mdi:infinity",
+      "state_class": "total_increasing",
+      "unit_of_measurement": "kWh"
+   },
+   "meter_self_eq": {
+      "name": "self_eq",
+      "device_class": "energy",
+      "icon": "mdi:infinity",
+      "state_class": "total_increasing",
+      "unit_of_measurement": "kWh"
+   },
+   "bms_return": {
+      "name": "bms_out_eq",
+      "device_class": "energy",
+      "icon": "mdi:infinity",
+      "state_class": "total_increasing",
+      "unit_of_measurement": "kWh"
+   },
+   "bms_charging": {
+      "name": "bms_in_eq",
+      "device_class": "energy",
+      "icon": "mdi:infinity",
+      "state_class": "total_increasing",
+      "unit_of_measurement": "kWh"
+   },
+   "bms_soc": {
+      "name": "bms_soc",
+      "device_class": "battery",
+      "icon": "mdi:infinity",
+      "unit_of_measurement": "%"
+   },
+   "todos": {
+      "name": "",
+      "device_class": "",
+      "state_class": "",
+      "icon": "",
+      "unit_of_measurement": ""
    }
 }
